### 2022-03-30 #523 @msperlin
yfR: Downloads and Organizes Financial Data from Yahoo Finance
Submitting Author Name: Marcelo Perlin
Submitting Author Github Handle: msperlin
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/msperlin/yfR
Version submitted: 0.0.1
Submission type: Standard

Version accepted: TBD
---

```
Package: yfR
Title: Downloads and Organizes Financial Data from Yahoo Finance
Version: 0.0.1
Authors@R: person("Marcelo", "Perlin", email = "marceloperlin@gmail.com", role = c("aut", "cre"))
Description: Facilitates download of financial data from Yahoo Finance <https://finance.yahoo.com/>,
 a vast repository of stock price data across multiple financial exchanges. The package offers a local caching system
 and support for parallel computation.
Imports: stringr, curl, tidyr,
    lubridate, furrr, purrr, future, tibble, zoo,
    cli, readr, rvest, dplyr, quantmod
    covr
```

Package yfR retrieves and organizes data from [Yahoo Finance](https://finance.yahoo.com/), a large repository for stock price data.

Target audience are students, researchers and industry practioneers in the field of Finance and Economics.

Package yfR is the second and backwards-incompatible version of [BatchGetSymbols](https://github.com/msperlin/BatchGetSymbols), also developed by me. My plan is to first deprecate BatchGetSymbols and later remove it from CRAN and archive it in Github.

Moreover, there are other packages, such as [quantmod](https://www.quantmod.com/), that downloads data from Yahoo Finance, but none with similar features to yfR and BatchGetSymbols.

Yes.

Unfortinately, I was not able to run pkgcheck locally as I was unable to install (or make) dependency ctags in my Linux Mint 20.3 machine. Nonetheless, I read through and followed all guidelines available in the [manual](https://devguide.ropensci.org/authors-guide.html).
