### 2022-03-31 30526f6
added R-cmd-check badge & rebuild docs
