+++ README.md
@@ -7,6 +7,9 @@ state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Codecov test
 coverage](https://codecov.io/gh/msperlin/yfR/branch/main/graph/badge.svg)](https://app.codecov.io/gh/msperlin/yfR?branch=main)
+[![R build
+(rcmdcheck)](https://github.com/msperlin/yfR/workflows/R-CMD-check/badge.svg)](https://github.com/msperlin/yfR/actions)
+
 <!-- badges: end -->
 
 # Motivation
@@ -130,15 +133,15 @@ df_yf <- yf_get(tickers = my_ticker,
                      first_date = first_date,
                      last_date = last_date)
 #> 
-#> ── Running yfR for 1 stocks | 2022-02-28 --> 2022-03-30 (30 days) ──
+#> ── Running yfR for 1 stocks | 2022-03-01 --> 2022-03-31 (30 days) ──
 #> 
 #> ℹ Downloading data for benchmark ticker ^GSPC
 #> ℹ (1/1) Fetching data for ',
 #> 'FB
 #> !    - not cached
 #> ✓    - cache saved successfully
-#> ✓    - got 22 valid rows (2022-02-28 --> 2022-03-29)
-#> ✓    - got 100% of valid prices -- Good job msperlin!
+#> ✓    - got 22 valid rows (2022-03-01 --> 2022-03-30)
+#> ✓    - got 100% of valid prices -- Good stuff!
 #> ℹ Binding price data
 
 # output is a tibble with data
@@ -146,12 +149,12 @@ head(df_yf)
 #> # A tibble: 6 × 10
 #>   ticker ref_date   price_open price_high price_low price_close   volume
 #>   <chr>  <date>          <dbl>      <dbl>     <dbl>       <dbl>    <dbl>
-#> 1 FB     2022-02-28       208.       213.      207.        211. 34239800
-#> 2 FB     2022-03-01       210.       212.      202.        203. 27094900
-#> 3 FB     2022-03-02       205.       209.      202.        208. 29452100
-#> 4 FB     2022-03-03       209.       209.      201.        203. 27263500
-#> 5 FB     2022-03-04       202.       206.      199.        200. 32130900
-#> 6 FB     2022-03-07       201.       201.      187.        187. 38560600
+#> 1 FB     2022-03-01       210.       212.      202.        203. 27094900
+#> 2 FB     2022-03-02       205.       209.      202.        208. 29452100
+#> 3 FB     2022-03-03       209.       209.      201.        203. 27263500
+#> 4 FB     2022-03-04       202.       206.      199.        200. 32130900
+#> 5 FB     2022-03-07       201.       201.      187.        187. 38560600
+#> 6 FB     2022-03-08       188.       197.      186.        190. 37508100
 #> # … with 3 more variables: price_adjusted <dbl>, ret_adjusted_prices <dbl>,
 #> #   ret_closing_prices <dbl>
 ```
@@ -170,26 +173,26 @@ df_yf_multiple <- yf_get(tickers = my_ticker,
                               first_date = first_date,
                               last_date = last_date)
 #> 
[…]
