### 2022-03-31 #523 @msperlin
Good morning.

Sure, I just added the R-CMD badge.
