### 2022-03-31 1ee2f6f
fixed typo in Readme.Rmd & rebuild
