### 2022-04-01 #523 @jooolia
Dear @msperlin,
Thank you for your submission. The package has passed all of the automated package checks and the test coverage is good.
Could you expand a bit more on how this package differs from quantmod and tidyquant?
Thanks, Julia
