### 2022-04-02 #523 @msperlin
Good morning Julia,

The main goal of yfR is to help user download large ammounts of data from Yahoo Finance (YF).

Packages quantmod and tidyquant also offers a function for downloading price data from YF, but only that. Besides importing data, yfR offers the following functionalities:

- Organization and clean up of data
    - Users can set a threshold for what is "bad" data with respect to matching dates to a benchmark dataset (SP500 is usually used);
    - Users can also ask for "complete data", where all missing dates are set as NA for later substitution;
    - Log or arithmetic returns, much used in research, are also calculated by default;
    - User can aggregate the data to weekly, monthly or yearly, always keeping the same data structure.

- smarter downloads
    - A local (and smart) session-persistent caching system is implemented. This means that, within a session, the data is never downloaded twice and only missing portions of data are downloaded;
    - Support for parallel computing. Users can easily set up concurrent R sessions for faster download of data.

- Practicality
    - yfR innovates with a  "collection"  system, where one can easily import a collection of tickers such as the SP500 composition in a single function call.
