### 2022-04-08 c6b21f1
added pkgdown bot build in github-actions
