+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://msperlin.github.io/yfR/
+template:
+  bootstrap: 5
+
