+++ pkgdown.yml
@@ -3,5 +3,8 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   yf-vignette: yf-vignette.html
-last_built: 2022-03-30T18:25Z
+last_built: 2022-04-08T12:46Z
+urls:
+  reference: https://msperlin.github.io/yfR/reference
+  article: https://msperlin.github.io/yfR/articles
 
