### 2022-04-08 b41efa6
fixex #4 (using magrittr's pipe for compatibility with R 4.0.0)
