+++ inst/scripts/S_create_logo.R
@@ -1,8 +1,8 @@
 library(hexSticker)
 library(yfR)
 
-df_sp500 <- yfR::yf_get('^GSPC', first_date = '1950-01-01') |>
-  dplyr::ungroup() |>
+df_sp500 <- yfR::yf_get('^GSPC', first_date = '1950-01-01') %>%
+  dplyr::ungroup() %>%
   dplyr::select(ref_date, price_adjusted)
 
 
