### 2022-04-08 466e199
fix #4 (using magrittr pipe) -- now with correct git message
