+++ vignettes/yf-vignette.R
@@ -7,9 +7,9 @@ first_date <- Sys.Date() - 30
 last_date <- Sys.Date()
 
 # fetch data
-df_yf <- yf_get(tickers = my_ticker, 
-                     first_date = first_date,
-                     last_date = last_date)
+df_yf <- yf_get(tickers = my_ticker,
+                first_date = first_date,
+                last_date = last_date)
 
 # output is a tibble with data
 head(df_yf)
@@ -22,13 +22,13 @@ my_ticker <- c('FB', 'GM', 'MMM')
 first_date <- Sys.Date() - 100
 last_date <- Sys.Date()
 
-df_yf_multiple <- yf_get(tickers = my_ticker, 
-                     first_date = first_date,
-                     last_date = last_date)
+df_yf_multiple <- yf_get(tickers = my_ticker,
+                         first_date = first_date,
+                         last_date = last_date)
 
 
 p <- ggplot(df_yf_multiple, aes(x = ref_date, y = price_adjusted,
-                       color = ticker)) + 
+                                color = ticker)) +
   geom_line()
 
 p
@@ -36,9 +36,9 @@ p
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  library(yfR)
-#  
+#
 #  df_yf <- yf_collection_get("SP500")
-#  
+#
 #  head(df_yf)
 
 ## -----------------------------------------------------------------------------
@@ -50,41 +50,41 @@ my_ticker <- 'GE'
 first_date <- '2010-01-01'
 last_date <- Sys.Date()
 
-df_dailly <- yf_get(tickers = my_ticker, 
-                         first_date, last_date, 
-                         freq_data = 'daily') %>%
+df_dailly <- yf_get(tickers = my_ticker,
+                    first_date, last_date,
+                    freq_data = 'daily') %>%
   mutate(freq = 'daily')
-  
-  
-df_weekly <- yf_get(tickers = my_ticker, 
-                         first_date, last_date, 
-                         freq_data = 'weekly') %>%
+
+
+df_weekly <- yf_get(tickers = my_ticker,
+                    first_date, last_date,
+                    freq_data = 'weekly') %>%
   mutate(freq = 'weekly')
 
-df_monthly <- yf_get(tickers = my_ticker, 
-                         first_date, last_date, 
-                         freq_data = 'monthly') %>%
+df_monthly <- yf_get(tickers = my_ticker,
+                     first_date, last_date,
+                     freq_data = 'monthly') %>%
   mutate(freq = 'monthly')
 
-df_yearly <- yf_get(tickers = my_ticker, 
-                         first_date, last_date, 
-                         freq_data = 'yearly') %>%
+df_yearly <- yf_get(tickers = my_ticker,
+                    first_date, last_date,
+                    freq_data = 'yearly') %>%
   mutate(freq = 'yearly')
 
 df_allfreq <- bind_rows(
   list(df_dailly, df_weekly, df_monthly, df_yearly)
 ) %>%
-  mutate(freq = factor(freq, 
-                       levels = c('daily', 
+  mutate(freq = factor(freq,
+                       levels = c('daily',
                                   'weekly',
                                   'monthly',
                                   'yearly'))) # make sure the order in plot is right
 
-p <- ggplot(df_allfreq, aes(x=ref_date, y = price_adjusted)) + 
-  geom_point() + 
-  geom_line() + 
-  facet_grid(freq ~ ticker) + 
-  theme_minimal() + 
+p <- ggplot(df_allfreq, aes(x=ref_date, y = price_adjusted)) +
+  geom_point() +
+  geom_line() +
[…]
