+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   yf-vignette: yf-vignette.html
-last_built: 2022-04-08T12:46Z
+last_built: 2022-04-08T12:55Z
 urls:
   reference: https://msperlin.github.io/yfR/reference
   article: https://msperlin.github.io/yfR/articles
