+++ DESCRIPTION
@@ -9,7 +9,7 @@ URL: https://github.com/msperlin/yfR,
     https://msperlin.github.io/yfR/
 BugReports: https://github.com/msperlin/yfR/issues
 Depends:
-    R (>= 4.1)
+    R (>= 4.0.0)
 Imports: 
     stringr,
     curl,
