### 2022-05-01 ca92e4d
fixes #5 (issue with api limits)
