+++ README.Rmd
@@ -25,7 +25,7 @@ knitr::opts_chunk$set(
 
 `yfR` is the second and backwards-incompatible version of [BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols). In a nutshell, it provides access to daily stock prices from [Yahoo Finance](https://finance.yahoo.com/), a vast repository with financial data around the globe. Yahoo Finance cover a large number of markets and assets, being used extensively for importing price datasets used in academic research and teaching. 
 
-Package `yfR` is based on [quantmod](https://www.quantmod.com/) and used its main function for fetching data from Yahoo Finance. The main innovation in `yfR` is in the organization of the imported financial data and using local caching system and parallel computing for speeding up large scale download of datasets from Yahoo Finance.
+Package `yfR` is based on [quantmod](https://www.quantmod.com/) and uses one of its functions (`quantmod::getSymbols`) for fetching data from Yahoo Finance. The main innovation in `yfR` is in the organization of the imported financial data and using a local caching system and parallel computing for speeding up large scale download of datasets from Yahoo Finance.
 
 See full documentation [here](https://github.com/msperlin/yfR).
 
@@ -50,13 +50,13 @@ Package `BatchgetSymbols` was developed back in 2016, with many bad structural c
 
 Here are the main differences between `yfR` (new) and `BatchGetSymbols` (old): 
 
-- All input arguments are now formatted as "snake_case" and not "dot.case". For example, the argument for the first date of data importation in `yfR::yf_get()` is `first_date`, and not `first.date` as used in  `BatchGetSymbols::BatchGetSymbols`.
+- All input arguments are now formatted as "snake_case" and not "dot.case". For example, the argument for the first date of data importation in `yfR::yf_get()` is `first_date`, and not `first.date`, as used in  `BatchGetSymbols::BatchGetSymbols`.
+
+- A new feature called "collection", which allows for easy download of a collection of tickers. For example, you can download price data for all components of the SP500 by simply calling `yfR::yf_collection_get("SP500")`.
 
 - All function have been renamed for a common API notation. For example, `BatchGetSymbols::BatchGetSymbols` is now `yfR::yf_get()`. Likewise, the function for fetching collections is `yfR::yf_collection_get()`.
 
 - The output of `yfR::yf_get()` is always a tibble with the price data (and not a list as in `BatchGetSymbols::BatchGetSymbols`). If one wants the tibble with a summary of the importing process, it is available as an attribute of the output (see function `base::attributes`)
-
-- A new feature called "collection", which allows for easy download of a collection of tickers. For example, you can download price data for all components of the SP500 by simply calling `yfR::yf_collection_get("SP500")`.
 
 - New and prettier status messages using package `cli`
 
