+++ README.md
@@ -22,11 +22,12 @@ data around the globe. Yahoo Finance cover a large number of markets and
 assets, being used extensively for importing price datasets used in
 academic research and teaching.
 
-Package `yfR` is based on [quantmod](https://www.quantmod.com/) and used
-its main function for fetching data from Yahoo Finance. The main
-innovation in `yfR` is in the organization of the imported financial
-data and using local caching system and parallel computing for speeding
-up large scale download of datasets from Yahoo Finance.
+Package `yfR` is based on [quantmod](https://www.quantmod.com/) and uses
+one of its functions (`quantmod::getSymbols`) for fetching data from
+Yahoo Finance. The main innovation in `yfR` is in the organization of
+the imported financial data and using a local caching system and
+parallel computing for speeding up large scale download of datasets from
+Yahoo Finance.
 
 See full documentation [here](https://github.com/msperlin/yfR).
 
@@ -71,8 +72,13 @@ Here are the main differences between `yfR` (new) and `BatchGetSymbols`
 
 -   All input arguments are now formatted as “snake_case” and not
     “dot.case”. For example, the argument for the first date of data
-    importation in `yfR::yf_get()` is `first_date`, and not `first.date`
-    as used in `BatchGetSymbols::BatchGetSymbols`.
+    importation in `yfR::yf_get()` is `first_date`, and not
+    `first.date`, as used in `BatchGetSymbols::BatchGetSymbols`.
+
+-   A new feature called “collection”, which allows for easy download of
+    a collection of tickers. For example, you can download price data
+    for all components of the SP500 by simply calling
+    `yfR::yf_collection_get("SP500")`.
 
 -   All function have been renamed for a common API notation. For
     example, `BatchGetSymbols::BatchGetSymbols` is now `yfR::yf_get()`.
@@ -84,11 +90,6 @@ Here are the main differences between `yfR` (new) and `BatchGetSymbols`
     wants the tibble with a summary of the importing process, it is
     available as an attribute of the output (see function
     `base::attributes`)
-
--   A new feature called “collection”, which allows for easy download of
-    a collection of tickers. For example, you can download price data
-    for all components of the SP500 by simply calling
-    `yfR::yf_collection_get("SP500")`.
 
 -   New and prettier status messages using package `cli`
 
@@ -133,15 +134,14 @@ df_yf <- yf_get(tickers = my_ticker,
                      first_date = first_date,
                      last_date = last_date)
 #> 
-#> ── Running yfR for 1 stocks | 2022-03-01 --> 2022-03-31 (30 days) ──
+#> ── Running yfR for 1 stocks | 2022-04-01 --> 2022-05-01 (30 days) ──
 #> 
 #> ℹ Downloading data for benchmark ticker ^GSPC
-#> ℹ (1/1) Fetching data for ',
-#> 'FB
+#> ℹ (1/1) Fetching data for FB
 #> !    - not cached
[…]
