+++ NEWS.md
@@ -1,3 +1,8 @@
+## Version 0.0.2 (2022-05-02)
+
+- Added warning message for parallel option (yahoo finance has just set a api limit)
+- Fixed typos in readme.md
+
 ## Version 0.0.1 (2022-03-28)
 
 - First version, ported from [BatchGetSymbols](https://github.com/msperlin/BatchGetSymbols)
