+++ README.md
@@ -9,7 +9,8 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 coverage](https://codecov.io/gh/msperlin/yfR/branch/main/graph/badge.svg)](https://app.codecov.io/gh/msperlin/yfR?branch=main)
 [![R build
 (rcmdcheck)](https://github.com/msperlin/yfR/workflows/R-CMD-check/badge.svg)](https://github.com/msperlin/yfR/actions)
-
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/523_status.svg)](https://github.com/ropensci/software-review/issues/523)
 <!-- badges: end -->
 
 # Motivation
@@ -134,14 +135,14 @@ df_yf <- yf_get(tickers = my_ticker,
                      first_date = first_date,
                      last_date = last_date)
 #> 
-#> ── Running yfR for 1 stocks | 2022-04-01 --> 2022-05-01 (30 days) ──
+#> ── Running yfR for 1 stocks | 2022-04-02 --> 2022-05-02 (30 days) ──
 #> 
 #> ℹ Downloading data for benchmark ticker ^GSPC
 #> ℹ (1/1) Fetching data for FB
 #> !    - not cached
 #> ✔    - cache saved successfully
-#> ✔    - got 20 valid rows (2022-04-01 --> 2022-04-29)
-#> ✔    - got 100% of valid prices -- Nice!
+#> ✔    - got 19 valid rows (2022-04-04 --> 2022-04-29)
+#> ✔    - got 100% of valid prices -- Good job msperlin!
 #> ℹ Binding price data
 
 # output is a tibble with data
@@ -149,12 +150,12 @@ head(df_yf)
 #> # A tibble: 6 × 10
 #>   ticker ref_date   price_open price_high price_low price_close   volume
 #>   <chr>  <date>          <dbl>      <dbl>     <dbl>       <dbl>    <dbl>
-#> 1 FB     2022-04-01       225.       227.      223.        225. 19533300
-#> 2 FB     2022-04-04       226.       234.      226.        234. 28054800
-#> 3 FB     2022-04-05       234.       237.      231.        232. 29727200
-#> 4 FB     2022-04-06       227.       229.      221.        223. 28995100
-#> 5 FB     2022-04-07       224.       226.      219.        223. 21037200
-#> 6 FB     2022-04-08       222.       225.      220.        222. 18363700
+#> 1 FB     2022-04-04       226.       234.      226.        234. 28054800
+#> 2 FB     2022-04-05       234.       237.      231.        232. 29727200
+#> 3 FB     2022-04-06       227.       229.      221.        223. 28995100
+#> 4 FB     2022-04-07       224.       226.      219.        223. 21037200
+#> 5 FB     2022-04-08       222.       225.      220.        222. 18363700
+#> 6 FB     2022-04-11       218.       221.      215.        216. 20516600
 #> # … with 3 more variables: price_adjusted <dbl>, ret_adjusted_prices <dbl>,
 #> #   ret_closing_prices <dbl>
 ```
@@ -173,24 +174,24 @@ df_yf_multiple <- yf_get(tickers = my_ticker,
                               first_date = first_date,
                               last_date = last_date)
 #> 
[…]
