+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: yfR
 Title: Downloads and Organizes Financial Data from Yahoo Finance
-Version: 0.0.1
+Version: 0.0.2
 Authors@R: person("Marcelo", "Perlin", email = "marceloperlin@gmail.com", role = c("aut", "cre"))
 Description: Facilitates download of financial data from Yahoo Finance <https://finance.yahoo.com/>, 
  a vast repository of stock price data across multiple financial exchanges. The package offers a local caching system
