### 2022-05-03 1b65cbd
added cumulative returns in output
