### 2022-05-05 #523 @Athene-ai
@melvidoni thanks for adding me as reviewer and I filled the volunteer form for being  an rOpenSci Reviewer :-)
