### 2022-05-09 2e236b8
now passing all arguments of yf_get() to yf_collections_get() [fixes #6]
