### 2022-05-09 #523 @melvidoni
> @melvidoni do we have a slack channel?

Hello @Athene-ai. Please, be mindful that responses are not immediate, _especially_ over the weekend; kindly do not hasten people, and wait for responses/actions. There is much going on "behind the scenes" that you may not be aware of.

That said, you'll get an invitation to the Slack later in the process.
