### 2022-05-10 #523 @Athene-ai
> > @melvidoni do we have a slack channel?
>
> Hello @Athene-ai. Please, be mindful that responses are not immediate, _especially_ over the weekend; kindly do not hasten people, and wait for responses/actions. There is much going on "behind the scenes" that you may not be aware of.
>
> That said, you'll get an invitation to the Slack later in the process.

Thanks for the information 😊
