### 2022-05-20 #523 @Athene-ai
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

-   [5] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

In general I think this effort is a great one. I like the idea of having a package able to provide access to daily stock prices from stock market trend such as here Yahoo Finance.It also supports increased usability and replicability of a project on the "left side" of the spectrum - i.e. getting the data initially. I support and applaud this effort!

However, I have a few general concerns that I think could be addressed. I'd like to better understand the use cases for this effort and the users technical knowledge. In general, most people don't fully understand how important these type of financial data are. And often the starting place is to "download everything" without recognizing that for instance their hard drive is quickly filling up. Thus, I think this package could provide some guidance for users before they download and attempt to archive everything.

For instance,

-   Perhaps there is some guidance associated with the types of data that might be appropriate to use with this package

-   Perhaps there are some guidelines and warning associated with downloading data and the space availability on the drive were the data are being downloaded.

I think some of these learning points are important to a successful package because so many users could quickly make huge API calls and fill up their hard drives unknowingly.

**Use Cases**

I'd like to see several clear uses cases for this package that include the user personas who are accessing the data . I see many potential challenges for users who are less well versed in working with these data - particularly in the data size realm and dealing with API calls and associated arguments.

I'd also like to see a bit more documentation and warnings associated with potentially large requests and associated resource availability on a users machine.

**Package Documentation**

In general, the readme and introductory package text is written for someone with more technical experience and who has probably used API's before even though the audience as described is "researchers". If the audience is anyone who is looking to access data and build metadata around it, I think more clear documentation is required up front about what the package does and how to use it. I provide some specifics in this regard below!

I also think that all of the function documentation needs to be developed further.

**Setup Feedback**

First I installed the package from github. Issues with the installation.

 Package installation requires a pre-requisite that are not present in the installation README.md section

**SETUP**

-   Add more explicit instructions and discussion about installation. Add instructions

    for both Mac and Windows and be explicitly about which install path is for which

    OS. If you expect to only support advanced users, then they may be able to figure it out.

\ Truly great effort and tremendous work has obviously been put into this package! Please take this feedback as an attempt to help with the eye of a regular (non power) user. :)

Also please let me know if I missed some key parts of this review.
