### 2022-05-22 565b3f9
now using json entrypoint at quantmod::getsymbols() -- fixes #9
