+++ NEWS.md
@@ -1,3 +1,12 @@
+## Version 0.0.4 (2022-05-22)
+
+- new call to quantmod::GetSymbols() using json entry point, which, so far, 
+is unrestricted. This fixes [issue 9](https://github.com/msperlin/yfR/issues/9).
+- re-inserted the parallel option in yf_get (passed all tests)
+- now using a function for testing collection download
+
+- Added cumulative return at output
+
 ## Version 0.0.3 (2022-05-03)
 
 - Added cumulative return at output
