+++ DESCRIPTION
@@ -28,7 +28,7 @@ Imports:
     magrittr
 License: MIT + file LICENSE
 LazyData: true
-RoxygenNote: 7.1.2
+RoxygenNote: 7.2.0
 Suggests: 
     knitr,
     rmarkdown,
