### 2022-05-29 #523 @s3alfisc
<table>
<tbody><tr class="odd"><td>title: “review”</td></tr><tr class="even"><td>output:</td></tr><tr class="odd"><td>rmarkdown::md_document:</td></tr><tr class="even"><td>pandoc_args: [</td></tr><tr class="odd"><td>“–wrap=none”</td></tr><tr class="even"><td>]</td></tr></tbody>
</table>

## Package Review

-   **Briefly describe any working relationship you have (had) with the package authors.** None
-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☐ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README. *While a readme and pkgdown website exists, I believe that the documentation could be greatly improved - see my comments below.*
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally *Vignette runs locally*.
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`). *All available.*

*You can find more comments on documentation below.*

#### Functionality

-   ☐ **Installation:** Installation succeeds as documented on my windows machine. *On github actions, the cmd check currently fails for mac.*
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
    *There are no performance claims made in the package.*
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions.
    *All tests pass on the local machine. But coverage is below 80% - I would love to see this go above 95% :)*
-   ☐ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.
    +  ☐ package name: maybe a search engine optimized name for the package would be `yfinanceR` - the name of the equivalent Python package is `yfinance`
    + ☐ create package metadata with `codemeta`
    + ☐ As per the CMD check failure for Mac, the package currently does not run for Mac.
    + ☒ functions have descriptive names; snake case; function argument order consistent across functions; no conflicts with base packages
    + ☒ console messages via `error()` and `warning()`
    + ☐ no citation file is included
    + ☐ (consider creating a r-universe profile & adding a r-universe badge, r-universe is great :) )
    + ☐ add installation instructions using `remotes`, `pak`, etc
    + ☒ code of conduct, contribution guidelines available
    + ☐ the package does not contain top-level documentation - i.e. `??yfR` does not return any documentation of the package
    + ☐ add a link to the readme with further extensive documentation on `yahoo finance
    + ☐ "If your package provides access to a data source, we require that DESCRIPTION contains both (1) A brief identification and/or description of the organisation responsible for issuing data; and (2) The URL linking to public-facing page providing, describing, or enabling data access (which may often differ from URL leading directly to data source)."
    + ☒ `roxygen2` is used
    + ☐ in general, `@return` statements specify the returned data object. But you could be more specific - usually, `tibbles` are returned, not base `data.frames
    + ☒ `@noRD` used for non-exported functions
    + ☒ `pkgdown` website exists
    + ☒ license: MIT
    + ☒ all user facing functions have examples
    + ☐ package dependencies: by running `pkgstats`, it looks like there are multiple package dependencies that you could easily replace by using `base` functions, e.g. `dplyr`, `magrittr`, `tibble`. What is the advantage of using `readr::read_rds()` over `base::readRDS()`? The packages `curl`, `cli`, is not detected in use by `pkgstats`. Nevertheless, all imported packages are of high quality, so I have no concerns here.

Estimated hours spent reviewing: 8

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Additional Comments

I think that `yfR` is a very promising package with useful features, and I believe that it will be widely used. I very much enjoyed using it! To improve the package, I mostly suggest to invest more time into refining the documentation.

#### Documentation

-   Statement of need: I would like to see a more refined statement of need at the beginning of the readme: what is `yfR`’s main innovation? E.g. start with something like “`yfR` is an API to yahoo finance. It speeds up the data downloading process by parallel computing and local caching.” Then explain what type of data yahoo finance includes.
-   I would move the discussion of data quality / limitations of yahoo finance and comparison to `BatchGetSymbols` to separate articles - I don’t think they are required in the readme.
-   If you want to keep the reference to `quantmod`, maybe include a dedicated ‘Acknowledgements’ section at the end of the readme?
-   Occasionally, you use jargon: e.g., not all users might now what a ticker is.
- I would move all examples from the readme to the ‘get started’ vignette. Alternatively, I would keep only one example in the readme.
-   In the ‘get started’ vignette, I would hide the message output generated e.g. by `yf_get()` and explain in words what the function does: e.g. it checks the cache, downloads data if the cache is empty, else finishes etc.
-   The vignette states that multiple ‘collections’ are organized in the package. It would be great to include a full list of collections to the docs, e.g. as a separate article? The `yf_get_available_collections()` helps here, but what do the individual collections stand for? E.g. does IBOV stand for the Bovespa-Index?
-   I would like to see some documentation on how the caching works: e.g., where are files saved? For how long are they saved? Is the cache ever cleaned, e.g. are cached files lost by re-starting the R session?
-   In the docs for `yf_convert_to_wide`, it would be good to print the initial long dataframe.
-   The documentation of `yf_get()` does not really, as a stand-alone, explain what the function does: download ticker data from yahoo finance, caching, parallelism etc. I would delete the reference to `getSymbols`. Note that as `yf_get_default_cache_folder()` is not exported, users will run into an error when trying `yfr::yf_get_default_cache_folder()`. Also, mention that the `ticker` function argument is vectorized
-   You could improve the documentation for parallelism: I myself have never used `furrr`, so your hint to `furrr::plan()` is not too helpful. How about a dedicated article with a small example that illustrates how to run `get_plan()` in parallel? Also, I only learned from browsing the code that by default, half of all available cores are used.
-   What is the difference between a collection and an index?
-   Consider adding documentation of the data returned via `yf_get()`. Not being a financial economist, I for example have no idea what the `price_adjusted` column stands for. Beyond, what is the unit of measurement of the price variables? I suppose it is US Dollars? Further, what is the relationship between daily data and monthly data? Also, potentially add a note that when markets are closed, no data row will be created.
- examples could be more 'verbose', i.e. add documentation
- also, examples could be more 'exhaustive' - they are quite minimal at the moment
- the example for `yf_convert_to_wide` currently calls internal data - could you not simply attach the data set or load it?

#### Installation, Local CMD Check & pkgcheck

- Installation and CMD check pass without problems
- I tried to run `pkgcheck`, but failed to get it to run. I suggest to run the `pkgcheck` action on github actions, at least for the time of the review.

#### Testing

-   Code Coverage is currently only at around 80% - I would love to see this up at 95%, if not 100 :)

#### Functionality

-   All examples work very nicely. Overall, it was a lot of fun using the package!
-   In general, the console output is very helpful and very pretty!
-   I am not sure if I would have default function arguments for `first_date()` and `last_date()`. If you want to keep it, I would change it from 15 days to one month.
-   `yf_convert_to_wide()` is super helpful - great idea to directly include it in the package!
-   Could the API be more permissive, e.g. accept dates with format `dd-mm-yyyy`?
-   When trying the “SP500” collection example, I ran into several ‘error in download’ errors. Still, the function finished eventually with ‘binding price data’. What exactly is going on here? Did the function eventually manage to fetch all tickers? If no, could there be a final message, e.g. ‘300/500 tickers successfully fetched. To fetch all others, do this …’.
-   I have seen that there is already a PR opened to alert users when they have reached the yahoo finance limit. This is would indeed be a great feature!

#### Additional Functionality

- It would be great to add further collections, e.g. NASDAQ, DAX, SP30, FAANG etc
- The equivalent Python package, [yfinance](https://pypi.org/project/yfinance/) offers a range of additional functionality, e.g. data on dividents, stock splits, and institutional investors. Do you plan to incorporate any of these into the package in the future?
- Currently, the cached files are saved in the `rds` file format via `readr::read_rds()`. There might be faster and/or more memory-friendly alternatives available. Have you considered adding a function argument that would allow users to store files e.g. in the `parquet` file format?
- Have you considered to integrate an `autoplot` functions to plot stock prices. `autoplot` would e.g. generate plots similarly to those created in the readme / vignette.
- Would it be possible to give an estimate of consumed memory of all cached files prior to a download? I would also consider to export `yf_get_default_cache_folder()` so that users are aware of the function and can easily check where `yfR` creates the cache.

#### Misc

- Do you need to export the `magrittr` pipe when using it internally?
-   I took a brief glance at the error messages, and most of them are clear and easy to understand. Maybe you could rephrase

      # check for NA
      if (any(is.na(tickers))) {
        my_msg <- paste0(
          "Found NA value in ticker vector.",
          "You need to remove it before running BatchGetSymbols."
        )
        stop(my_msg)
      }

        if (class(first_date) != "Date") {
        stop("ERROR: cant change class of first_date to 'Date'")
      }

In general, I really like the [dreamerr](https://github.com/lrberge/dreamerr) package for function input type checks. [checkmate](https://github.com/mllg/checkmate) seems to be very popular, too.

With `dreamerr`, you could e.g. write

      # check threshold
      if ((thresh_bad_data < 0) | (thresh_bad_data > 1)) {
        stop("Input thresh_bad_data should be a proportion between 0 and 1")
      }

as

    dreamerr::check_arg(thresh_bad_data, "scalar numeric GT{0} LT{1}")

I can’t really follow this error message:

      if (!flag) {
        warning(stringr::str_glue(
          "\nIt seems you are using a non-default cache folder at {cache_folder}. ",
          "Be aware that if any stock event -- split or dividend -- happens ",
          "in between cache files, the resulting aggregate cache data will not ",
          "correspond to reality as some part of the price data will not be ",
          "adjusted to the event. For safety and reproducibility, my suggestion ",
          "is to use cache system only for the current session with tempdir(), ",
          "which is the default option."
        ))
      }

-   The collections are created via hard coded (wikipedia) URLs. This is likely prone to errors - what if e.g. the URLs change? I understand the attractiveness of this ‘dynamic’ lookup, as e.g. the composition of stock indices might change over time. Maybe you could add a second look-up link (in case the main URL breaks), or you could add a ‘fallback’ data.frame containing the names of all firms included in an index at a fixed date to fall back to?

- My last comment (repeating something I mentioned above): the equivalent python package is called `yfinance`. Maybe a better / SEO optimized name for the package would be `yfinanceR`?
