### 2022-05-30 #523 @msperlin
Thanks @s3alfisc  for the review! Appreciate it. Good ideas there.

I'll reply to all your comments in the next couple of days.
