+++ README.Rmd
@@ -23,9 +23,13 @@ knitr::opts_chunk$set(
 
 # Motivation
 
-`yfR` is the second and backwards-incompatible version of [BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols). In a nutshell, it provides access to daily stock prices from [Yahoo Finance](https://finance.yahoo.com/), a vast repository with financial data around the globe. Yahoo Finance cover a large number of markets and assets, being used extensively for importing price datasets used in academic research and teaching. 
+`yfR` facilitates importing data from Yahoo finance by organizing the data in the `tidy` format and speeding up the process using a cache system and parallel computing. `yfR` is the second and backwards-incompatible version of [BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols), released in 2016 (see vignette "yfR and BatchGetSymbols"). 
 
-Package `yfR` is based on [quantmod](https://www.quantmod.com/) and uses one of its functions (`quantmod::getSymbols`) for fetching data from Yahoo Finance. The main innovation in `yfR` is in the organization of the imported financial data and using a local caching system and parallel computing for speeding up large scale download of datasets from Yahoo Finance.
+In a nutshell, [Yahoo Finance](https://finance.yahoo.com/) provides a vast repository of stock price data around the globe. It cover a signficant number of markets and assets, being used extensively in academic research and teaching. In order to import the financial data, all you need is a ticker (id of a stock, e.g. "FB" for [Meta platforms](https://finance.yahoo.com/quote/FB?p=FB&.tsrc=fin-srch)) and a time period. 
+
+# Finding ticker
+
+The easiest way to find the tickers of a company stock is to search for it in [Yahoo Finance](https://finance.yahoo.com/) website. 
 
 See full documentation [here](https://github.com/msperlin/yfR).
 
@@ -44,22 +48,6 @@ See full documentation [here](https://github.com/msperlin/yfR).
 - Parallel computing with package `furrr` is available, speeding up the data importation process.
 
 
-## Differences from [BatchGetSymbols](https://github.com/msperlin/BatchGetSymbols)
-
-Package `BatchgetSymbols` was developed back in 2016, with many bad structural choices from my part. Since then, I learned more about R and its ecosystem, resulting in better and more maintainable code. However, it is impossible to keep compatibility with the changes I wanted to make, which is why I decided to develop a new (and fresh) package. 
-
-Here are the main differences between `yfR` (new) and `BatchGetSymbols` (old): 
-
-- All input arguments are now formatted as "snake_case" and not "dot.case". For example, the argument for the first date of data importation in `yfR::yf_get()` is `first_date`, and not `first.date`, as used in  `BatchGetSymbols::BatchGetSymbols`.
-
-- A new feature called "collection", which allows for easy download of a collection of tickers. For example, you can download price data for all components of the SP500 by simply calling `yfR::yf_collection_get("SP500")`.
-
-- All function have been renamed for a common API notation. For example, `BatchGetSymbols::BatchGetSymbols` is now `yfR::yf_get()`. Likewise, the function for fetching collections is `yfR::yf_collection_get()`.
-
[…]
