+++ README.md
@@ -15,22 +15,32 @@ Review](https://badges.ropensci.org/523_status.svg)](https://github.com/ropensci
 
 # Motivation
 
-`yfR` is the second and backwards-incompatible version of
-[BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols).
-In a nutshell, it provides access to daily stock prices from [Yahoo
-Finance](https://finance.yahoo.com/), a vast repository with financial
-data around the globe. Yahoo Finance cover a large number of markets and
-assets, being used extensively for importing price datasets used in
-academic research and teaching.
+`yfR` facilitates importing data from Yahoo finance by organizing the
+data in the `tidy` format and speeding up the process using a cache
+system and parallel computing. `yfR` is the second and
+backwards-incompatible version of
+[BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols),
+released in 2016 (see vignette “yfR and BatchGetSymbols”).
 
-Package `yfR` is based on [quantmod](https://www.quantmod.com/) and uses
-one of its functions (`quantmod::getSymbols`) for fetching data from
-Yahoo Finance. The main innovation in `yfR` is in the organization of
-the imported financial data and using a local caching system and
-parallel computing for speeding up large scale download of datasets from
-Yahoo Finance.
+In a nutshell, [Yahoo Finance](https://finance.yahoo.com/) provides a
+vast repository of stock price data around the globe. It cover a
+signficant number of markets and assets, being used extensively in
+academic research and teaching. In order to import the financial data,
+all you need is a ticker (id of a stock, e.g. “FB” for [Meta
+platforms](https://finance.yahoo.com/quote/FB?p=FB&.tsrc=fin-srch)) and
+a time period.
 
-See full documentation [here](https://github.com/msperlin/yfR).
+# Finding tickers
+
+The easiest way to find the tickers of a company stock is to search for
+it in [Yahoo Finance’s](https://finance.yahoo.com/) website. At the top
+page you’ll find a search bar.
+
+![YF Search](search-yf.png?raw=true "Example of search in YF") From
+there, you’ll that a company can have many different stocks traded at
+different markets. As the example shows, Petrobras is traded at NYQ (New
+York Exchange), SAO (Sao Paulo/Brazil - B3 exchange) and BUE (Buenos
+Aires/Argentina Exchange), all with different symbols (tickers).
 
 ## Features
 
@@ -60,40 +70,6 @@ See full documentation [here](https://github.com/msperlin/yfR).
 -   Parallel computing with package `furrr` is available, speeding up
     the data importation process.
 
-## Differences from [BatchGetSymbols](https://github.com/msperlin/BatchGetSymbols)
-
-Package `BatchgetSymbols` was developed back in 2016, with many bad
-structural choices from my part. Since then, I learned more about R and
-its ecosystem, resulting in better and more maintainable code. However,
-it is impossible to keep compatibility with the changes I wanted to
-make, which is why I decided to develop a new (and fresh) package.
-
-Here are the main differences between `yfR` (new) and `BatchGetSymbols`
-(old):
-
[…]
