+++ README.Rmd
@@ -33,46 +33,36 @@ The main function of the package, `yfR::yf_get`, returns a dataframe with the fi
 
 The return dataframe contains the following columns:
 
-ticker
-: The requested tickers (ids of stocks)
+**ticker**: The requested tickers (ids of stocks)
 
-ref_date
-: The reference day (this can also be year/month/week when using argument freq_data)
+**ref_date**: The reference day (this can also be year/month/week when using argument freq_data)
 
-price_open
-: The opening price of the day/period
+**price_open**: The opening price of the day/period
 
-price_high
-: The highest price of the day/period
+**price_high**: The highest price of the day/period
 
-price_close
-: The close/last price of the day/period
+**price_close**: The close/last price of the day/period
 
-volume
-: The financial volume of the day/period
+**volume**: The financial volume of the day/period
 
-price_adjusted
-: The stock price adjusted for corporate events such as splits, dividends and 
+**price_adjusted**: The stock price adjusted for corporate events such as splits, dividends and 
 others -- this is usually what you want/need for studying stocks as it 
 represents the actual financial performance of stockholders
 
-ret_adjusted_prices
-: The arithmetic or log return (see input type_return) for the adjusted stock 
+**ret_adjusted_prices**: The arithmetic or log return (see input type_return) for the adjusted stock 
 prices
 
-ret_adjusted_prices
-: The arithmetic or log return (see input type_return) for the closing stock 
+**ret_adjusted_prices**: The arithmetic or log return (see input type_return) for the closing stock 
 prices
 
-cumret_adjusted_prices
-: The accumulated arithmetic/log return for the period (starts at 100%)
+**cumret_adjusted_prices**: The accumulated arithmetic/log return for the period (starts at 100%)
 
 
 # Finding tickers
 
 The easiest way to find the tickers of a company stock is to search for it in [Yahoo Finance's](https://finance.yahoo.com/) website. At the top page you'll find a search bar:
 
-![YF Search](search-yf.png?raw=true "Example of search in YF")
+![YF Search](/inst/figures/search-yf.png?raw=true "Example of search in YF")
 
 From there, you'll that a company can have many different stocks traded at different markets. As the example shows, Petrobras is traded at NYQ (New York Exchange), SAO (Sao Paulo/Brazil - B3 exchange) and BUE (Buenos Aires/Argentina Exchange), all with different symbols (tickers). 
 
