+++ pkgdown.yml
@@ -2,8 +2,10 @@ pandoc: 2.14.2
 pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
-  yf-vignette: yf-vignette.html
-last_built: 2022-05-09T12:06Z
+  diff-batchgetsymbols: diff-batchgetsymbols.html
+  getting-started: getting-started.html
+  using-collections: using-collections.html
+last_built: 2022-06-01T18:15Z
 urls:
   reference: https://msperlin.github.io/yfR/reference
   article: https://msperlin.github.io/yfR/articles
