+++ README.md
@@ -135,29 +135,29 @@ df_yf <- yf_get(tickers = my_ticker,
                      first_date = first_date,
                      last_date = last_date)
 #> 
-#> ── Running yfR for 1 stocks | 2022-04-02 --> 2022-05-02 (30 days) ──
+#> ── Running yfR for 1 stocks | 2022-05-02 --> 2022-06-01 (30 days) ──
 #> 
 #> ℹ Downloading data for benchmark ticker ^GSPC
 #> ℹ (1/1) Fetching data for FB
 #> !    - not cached
 #> ✔    - cache saved successfully
-#> ✔    - got 19 valid rows (2022-04-04 --> 2022-04-29)
-#> ✔    - got 100% of valid prices -- Good job msperlin!
+#> ✔    - got 21 valid rows (2022-05-02 --> 2022-05-31)
+#> ✔    - got 100% of valid prices -- Nice!
 #> ℹ Binding price data
 
 # output is a tibble with data
 head(df_yf)
-#> # A tibble: 6 × 10
+#> # A tibble: 6 × 11
 #>   ticker ref_date   price_open price_high price_low price_close   volume
 #>   <chr>  <date>          <dbl>      <dbl>     <dbl>       <dbl>    <dbl>
-#> 1 FB     2022-04-04       226.       234.      226.        234. 28054800
-#> 2 FB     2022-04-05       234.       237.      231.        232. 29727200
-#> 3 FB     2022-04-06       227.       229.      221.        223. 28995100
-#> 4 FB     2022-04-07       224.       226.      219.        223. 21037200
-#> 5 FB     2022-04-08       222.       225.      220.        222. 18363700
-#> 6 FB     2022-04-11       218.       221.      215.        216. 20516600
-#> # … with 3 more variables: price_adjusted <dbl>, ret_adjusted_prices <dbl>,
-#> #   ret_closing_prices <dbl>
+#> 1 FB     2022-05-02       201.       212.      201.        211. 49915300
+#> 2 FB     2022-05-03       210.       215.      208.        212. 41556300
+#> 3 FB     2022-05-04       211.       224.      207.        223. 41375900
+#> 4 FB     2022-05-05       219.       220.      206.        208. 41129200
+#> 5 FB     2022-05-06       207.       209.      201.        204. 34733600
+#> 6 FB     2022-05-09       200.       203.      196.        196. 36303200
+#> # … with 4 more variables: price_adjusted <dbl>, ret_adjusted_prices <dbl>,
+#> #   ret_closing_prices <dbl>, cumret_adjusted_prices <dbl>
 ```
 
 ### Fetching many stock prices
@@ -174,24 +174,24 @@ df_yf_multiple <- yf_get(tickers = my_ticker,
                               first_date = first_date,
                               last_date = last_date)
 #> 
-#> ── Running yfR for 3 stocks | 2022-01-22 --> 2022-05-02 (100 days) ──
+#> ── Running yfR for 3 stocks | 2022-02-21 --> 2022-06-01 (100 days) ──
 #> 
 #> ℹ Downloading data for benchmark ticker ^GSPC
 #> ℹ (1/3) Fetching data for FB
-#> ✔    - found cache file (2022-04-04 --> 2022-04-29)
+#> ✔    - found cache file (2022-05-02 --> 2022-05-31)
[…]
