+++ README.md
@@ -40,40 +40,32 @@ in dollars, while price data for PETR3.SA (B3/BR) is measured in Reais
 
 The return dataframe contains the following columns:
 
-ticker  
-The requested tickers (ids of stocks)
+**ticker**: The requested tickers (ids of stocks)
 
-ref_date  
-The reference day (this can also be year/month/week when using argument
-freq_data)
+**ref_date**: The reference day (this can also be year/month/week when
+using argument freq_data)
 
-price_open  
-The opening price of the day/period
+**price_open**: The opening price of the day/period
 
-price_high  
-The highest price of the day/period
+**price_high**: The highest price of the day/period
 
-price_close  
-The close/last price of the day/period
+**price_close**: The close/last price of the day/period
 
-volume  
-The financial volume of the day/period
+**volume**: The financial volume of the day/period
 
-price_adjusted  
-The stock price adjusted for corporate events such as splits, dividends
-and others – this is usually what you want/need for studying stocks as
-it represents the actual financial performance of stockholders
+**price_adjusted**: The stock price adjusted for corporate events such
+as splits, dividends and others – this is usually what you want/need for
+studying stocks as it represents the actual financial performance of
+stockholders
 
-ret_adjusted_prices  
-The arithmetic or log return (see input type_return) for the adjusted
-stock prices
+**ret_adjusted_prices**: The arithmetic or log return (see input
+type_return) for the adjusted stock prices
 
-ret_adjusted_prices  
-The arithmetic or log return (see input type_return) for the closing
-stock prices
+**ret_adjusted_prices**: The arithmetic or log return (see input
+type_return) for the closing stock prices
 
-cumret_adjusted_prices  
-The accumulated arithmetic/log return for the period (starts at 100%)
+**cumret_adjusted_prices**: The accumulated arithmetic/log return for
+the period (starts at 100%)
 
 # Finding tickers
 
@@ -81,7 +73,8 @@ The easiest way to find the tickers of a company stock is to search for
 it in [Yahoo Finance’s](https://finance.yahoo.com/) website. At the top
 page you’ll find a search bar:
 
-![YF Search](search-yf.png?raw=true "Example of search in YF")
+![YF
+Search](/inst/figures/search-yf.png?raw=true "Example of search in YF")
 
 From there, you’ll that a company can have many different stocks traded
 at different markets. As the example shows, Petrobras is traded at NYQ
@@ -162,16 +155,12 @@ df_yf <- yf_get(tickers = my_ticker,
 #> !    - not cached
 #> ✔    - cache saved successfully
 #> ✔    - got 21 valid rows (2022-05-02 --> 2022-05-31)
-<<<<<<< HEAD
-#> ✔    - got 100% of valid prices -- Nice!
-=======
-#> ✔    - got 100% of valid prices -- Time for some tea?
->>>>>>> 96f43b84423481ad71b0dd96012f3b2af6b2e857
+#> ✔    - got 100% of valid prices -- Got it!
 #> ℹ Binding price data
 #> 
 #> ── Diagnostics ─────────────────────────────────────────────────────────────────
[…]
