+++ vignettes/getting-started.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "01-Getting Started"
+title: "Getting Started"
 author: "Marcelo Perlin"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{01-Getting Started}
+  %\VignetteIndexEntry{Getting Started}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
