+++ vignettes/diff-batchgetsymbols.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "09-yfR and BatchGetSymbols"
+title: "yfR and BatchGetSymbols"
 author: "Marcelo Perlin"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{09-yfR and BatchGetSymbols}
+  %\VignetteIndexEntry{yfR and BatchGetSymbols}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
