+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   diff-batchgetsymbols: diff-batchgetsymbols.html
   getting-started: getting-started.html
   using-collections: using-collections.html
-last_built: 2022-06-01T18:27Z
+last_built: 2022-06-01T18:30Z
 urls:
   reference: https://msperlin.github.io/yfR/reference
   article: https://msperlin.github.io/yfR/articles
