### 2022-06-04 #523 @thisisnic
I am currently working on my review of this package, and hope to finish it in the next few days if nothing unexpected comes up!  I had an issue when I was running the examples in the vignette though, and so to deliver partial feedback which might be useful in the meantime, I've opened this issue relating to it on the project repo: https://github.com/msperlin/yfR/issues/11
