### 2022-06-05 #523 @melvidoni
> I am currently working on my review of this package, and hope to finish it in the next few days if nothing unexpected comes up! I had an issue when I was running the examples in the vignette though, and so to deliver partial feedback which might be useful in the meantime, I've opened this issue relating to it on the project repo: [msperlin/yfR#11](https://github.com/msperlin/yfR/issues/11)

Hello Nicola, that's great, thank you!
