### 2022-06-06 msperlin/yfR#11 @msperlin
Hi @thisisnic, I tested here (Linux Mint/ubuntu) and its working fine. Also, github-actions also tested in Windows/MAc and it is also working there.

Perhaps you reached the YF call limit? If so, it releases the block in a couple of minutes. Can you try it again ?
