### 2022-06-06 msperlin/yfR#11 @msperlin
btw, did  you upgrade your packages in the last month? there was a change in quantmod which affected yfR.
