### 2022-06-06 9661960
added list of available collections at vignette
