+++ vignettes/using-collections.Rmd
@@ -16,7 +16,11 @@ Collections are just a bundle of tickers pre-organized in the package. For examp
 The available collections are:
 
 ```{r}
-yfR::yf_get_available_collections()
+available_collections <- yfR::yf_get_available_collections(
+  print_description = TRUE
+  )
+
+available_collections
 ```
 
 One can download the composition of the collection with `yf_collection_get`:
@@ -24,6 +28,7 @@ One can download the composition of the collection with `yf_collection_get`:
 ```{r, eval=FALSE}
 library(yfR)
 
+# be patient, it takes a while
 df_yf <- yf_collection_get("SP500")
 
 head(df_yf)
