### 2022-06-06 #523 @msperlin
Thanks @thisisnic,

No rush for the review. I'm also on holidays and might take a while to reply here.

As for the issue, I think you might have a problem with quantmod. Please update with latest version in CRAN and try again.
