+++ README.Rmd
@@ -23,39 +23,43 @@ knitr::opts_chunk$set(
 
 # Motivation
 
-`yfR` facilitates importing stock prices from Yahoo finance, organizing the data in the `tidy` format and speeding up the process using a cache system and parallel computing. `yfR` is the second and backwards-incompatible version of [BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols), released in 2016 (see vignette "99-yfR and BatchGetSymbols"). 
+`yfR` facilitates importing stock prices from Yahoo finance, organizing the data in the `tidy` format and speeding up the process using a cache system and parallel computing. `yfR` is the second and backwards-incompatible version of [BatchGetSymbols](https://CRAN.R-project.org/package=BatchGetSymbols), released in 2016 (see vignette [yfR and BatchGetSymbols](https://msperlin.github.io/yfR/articles/diff-batchgetsymbols.html) for details). 
 
-In a nutshell, [Yahoo Finance](https://finance.yahoo.com/) provides a vast repository of stock price data around the globe. It cover a signficant number of markets and assets, being used extensively in academic research and teaching. In order to import the financial data, all you need is a ticker (id of a stock, e.g. "GM" for [General Motors](https://finance.yahoo.com/quote/GM?p=GM&.tsrc=fin-srch)) and a time period (first and last date). 
+In a nutshell, [Yahoo Finance (YF)](https://finance.yahoo.com/) provides a vast repository of stock price data around the globe. It covers a signficant number of markets and assets, being used extensively in academic research and teaching. In order to import the financial data from YF, all you need is a ticker (id of a stock, e.g. "GM" for [General Motors](https://finance.yahoo.com/quote/GM?p=GM&.tsrc=fin-srch)) and a time period -- first and last date. 
+
 
 # The Data
 
-The main function of the package, `yfR::yf_get`, returns a dataframe with the financial data. All price data is measured at the unit of the financial exchange. For example, price data for FB (NYSE/US) is measures in dollars, while price data for PETR3.SA (B3/BR) is measured in Reais (Brazilian currency).
+The main function of the package, `yfR::yf_get`, returns a dataframe with the 
+financial data. All price data is measured at the unit of the financial 
+exchange. For example, price data for GM (NASDAQ/US) is measured in dollars, 
+while price data for PETR3.SA (B3/BR) is measured in Reais (Brazilian currency).
 
-The return dataframe contains the following columns:
+The returned data contains the following columns:
 
-**ticker**: The requested tickers (ids of stocks)
+**ticker**: The requested tickers (ids of stocks);
 
-**ref_date**: The reference day (this can also be year/month/week when using argument freq_data)
+**ref_date**: The reference day (this can also be year/month/week when using argument freq_data);
 
-**price_open**: The opening price of the day/period
+**price_open**: The opening price of the day/period;
 
-**price_high**: The highest price of the day/period
+**price_high**: The highest price of the day/period;
 
-**price_close**: The close/last price of the day/period
+**price_close**: The close/last price of the day/period;
 
-**volume**: The financial volume of the day/period
+**volume**: The financial volume of the day/period, in the unit of the exchange;
 
-**price_adjusted**: The stock price adjusted for corporate events such as splits, dividends and 
-others -- this is usually what you want/need for studying stocks as it 
-represents the actual financial performance of stockholders
+**price_adjusted**: The stock price adjusted for corporate events such as 
+splits, dividends and others -- this is usually what you want/need for studying 
+stocks as it represents the real financial performance of stockholders;
 
 **ret_adjusted_prices**: The arithmetic or log return (see input type_return) for the adjusted stock 
-prices
+prices;
 
[…]
