### 2022-06-06 msperlin/yfR#11 @thisisnic
Hi @msperlin, I tried multiple times on separate dates but without any previous calls to the API, and it failed both times with the same error.  Unfortunately, I can't try it again, as it only failed before I upgraded my packages, so it works fine now.  The 2 different outputs above show a change in quantmod version, so I think you're right - that could well be the package which needs the version number adding in the DESCRIPTION file.
