### 2022-06-06 #523 @thisisnic
Thanks @msperlin - no problem!  Great, that fixes it for me.  Would it be worth perhaps adding the version to the DESCRIPTION file as described here[1] in order to prevent other people with older package versions wondering why it's not working?

[1] https://r-pkgs.org/description.html#minimum-versions
