### 2022-06-07 2749f6f
added quantmod (>= 0.4.20) -- fixes #11
