### 2022-06-07 msperlin/yfR#11 @msperlin
Thanks. I added the minimum verstion of quantmod to the DESCRIPTION file.
