+++ R/yf_data_get_raw.R
@@ -1,8 +1,8 @@
 #' Get clean data from yahoo/google using quantmod::getSymbols
 #' @noRd
-yf_get_clean_data <- function(ticker,
-                              first_date,
-                              last_date) {
+yf_data_get_raw <- function(ticker,
+                            first_date,
+                            last_date) {
 
 
   # dont push luck with yahoo servers
@@ -24,40 +24,9 @@ yf_get_clean_data <- function(ticker,
       },
       silent = TRUE)
 
-    }) })
+    })
+  })
 
-  # PREVIOUS code using json
-
-  # my_cols <- readr::cols(
-  #   Date = readr::col_date(format = ""),
-  #   Open = readr::col_double(),
-  #   High = readr::col_double(),
-  #   Low = readr::col_double(),
-  #   Close = readr::col_double(),
-  #   `Adj Close` = readr::col_double(),
-  #   Volume = readr::col_double()
-  # )
-  #
-  # suppressWarnings({
-  #   try(
-  #     {
-  #       df_raw <- readr::read_csv(yf_csv_link, col_types = my_cols) %>%
-  #       dplyr::mutate(ticker = ticker) %>%
-  #       dplyr::rename(
-  #         ref_date = Date,
-  #         price_open = Open,
-  #         price_high = High,
-  #         price_low = Low,
-  #         price_close = Close,
-  #         price_adjusted = `Adj Close`,
-  #         volume = Volume
-  #       ) %>%
-  #       dplyr::arrange(ref_date) %>% # make sure dates are sorted,
-  #       dplyr::relocate(ticker, ref_date)
-  #     },
-  #     silent = T
-  #   )
-  # })
 
   # in case of error, return empty df
   if (nrow(df_raw) == 0) return(df_raw)
@@ -68,11 +37,11 @@ yf_get_clean_data <- function(ticker,
     dplyr::mutate(ref_date = ref_date,
                   ticker = ticker)
   #%>%
-   # as.data.frame(df_raw[!duplicated(zoo::index(df_raw))])
+  # as.data.frame(df_raw[!duplicated(zoo::index(df_raw))])
 
   colnames(df_raw) <- c('price_open','price_high','price_low',
-                     'price_close','volume','price_adjusted',
-                     'ref_date', 'ticker')
+                        'price_close','volume','price_adjusted',
+                        'ref_date', 'ticker')
 
   # further organization
   df_raw <- df_raw %>%
@@ -83,7 +52,7 @@ yf_get_clean_data <- function(ticker,
   # sometimes, yf outputs two data points for the same date (not sure why)
   df_raw <- df_raw %>%
     dplyr::group_by(ref_date, ticker) %>%
-    dplyr::filter(dplyr::row_number()==1)
+    dplyr::filter(dplyr::row_number() == 1)
 
   # make sure only unique rows are returned
   df_raw <- unique(df_raw)
