+++ man/yf_get_available_collections.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/collections.R
+% Please edit documentation in R/yf_collections.R
 \name{yf_get_available_collections}
 \alias{yf_get_available_collections}
 \title{Returns available collections}
@@ -8,7 +8,7 @@ yf_get_available_collections(print_description = FALSE)
 }
 \arguments{
 \item{print_description}{Logical (TRUE/FALSE) - flag for printing description of
-available colletions}
+available indices/collections}
 }
 \value{
 A string vector with available collections

[…]
