+++ man/yf_get_available_collections.Rd
@@ -8,7 +8,7 @@ yf_get_available_collections(print_description = FALSE)
 }
 \arguments{
 \item{print_description}{Logical (TRUE/FALSE) - flag for printing description of
-available colletions}
+available collections}
 }
 \value{
 A string vector with available collections
