### 2022-06-16 #523 @melvidoni
@thisisnic and @s3alfisc what are your thoughts on the changes?
Do let me know if you are satisfied with these changes, or if there is anything else outstanding.
