### 2022-06-21 #523 @s3alfisc
Hi all, I am very happy with the changes. The package looks and works great! In particular, I am super happy about how the documentation has evolved! :)
