### 2022-06-22 #523 @msperlin
Thanks @s3alfisc, @thisisnic! Realy appreciate the feedback.

I'll make all changes and report here soon.
