+++ DESCRIPTION
@@ -1,12 +1,12 @@
 Package: yfR
 Title: Downloads and Organizes Financial Data from Yahoo Finance
-Version: 0.0.5
+Version: 1.0.0
 Authors@R: person("Marcelo", "Perlin", email = "marceloperlin@gmail.com", role = c("aut", "cre"))
 Description: Facilitates download of financial data from Yahoo Finance <https://finance.yahoo.com/>, 
  a vast repository of stock price data across multiple financial exchanges. The package offers a local caching system
  and support for parallel computation.
-URL: https://github.com/msperlin/yfR,
-    https://msperlin.github.io/yfR/
+URL: https://github.com/ropensci/yfR,
+    https://docs.ropensci.org/yfR
 BugReports: https://github.com/msperlin/yfR/issues
 Depends:
     R (>= 4.0.0)
