### 2022-06-22 74f0528
fixed wordlist
