### 2022-06-22 #523 @msperlin
@melvidoni I've made all required changes. Please let me know if anything else is needed.
I'm also interested in a blog post @ropensci/blog-editors
