+++ DESCRIPTION
@@ -12,8 +12,8 @@ Description: Facilitates download of financial data from Yahoo Finance <https://
  a vast repository of stock price data across multiple financial exchanges. The package offers a local caching system
  and support for parallel computation.
 URL: https://github.com/ropensci/yfR,
-    https://docs.ropensci.org/yfR
-BugReports: https://github.com/msperlin/yfR/issues
+    https://docs.ropensci.org/yfR/
+BugReports: https://github.com/ropensci/yfR/issues
 Depends:
     R (>= 4.0.0)
 Imports: 
@@ -32,7 +32,8 @@ Imports:
     dplyr,
     quantmod (>= 0.4.20),
     magrittr,
-    humanize
+    humanize,
+    methods
 License: MIT + file LICENSE
 LazyData: true
 RoxygenNote: 7.2.0
