+++ R/yf_collections.R
@@ -14,7 +14,7 @@
 #'
 #' @examples
 #'
-#' \dontrun{
+#' \donttest{
 #' df_yf <- yf_collection_get(collection = "IBOV",
 #'                            first_date = Sys.Date() - 30,
 #'                            last_date = Sys.Date()
