### 2022-08-15 9eb93fe
fixed bug with FB ticker (CHECK failed)
