+++ vignettes/getting-started.Rmd
@@ -33,7 +33,7 @@ The steps of the algorithm are:
 library(yfR)
 
 # set options for algorithm
-my_ticker <- 'FB'
+my_ticker <- 'GM'
 first_date <- Sys.Date() - 30
 last_date <- Sys.Date()
 
@@ -53,7 +53,7 @@ head(df_yf)
 library(yfR)
 library(ggplot2)
 
-my_ticker <- c('FB', 'GM', 'MMM')
+my_ticker <- c('TSLA', 'GM', 'MMM')
 first_date <- Sys.Date() - 100
 last_date <- Sys.Date()
 
@@ -67,7 +67,6 @@ p <- ggplot(df_yf_multiple, aes(x = ref_date, y = price_adjusted,
   geom_line()
 
 p
-
 ```
 
 
@@ -128,7 +127,7 @@ print(p)
 library(yfR)
 library(ggplot2)
 
-my_ticker <- c('FB', 'GM', 'MMM')
+my_ticker <- c('TSLA', 'GM', 'MMM')
 first_date <- Sys.Date() - 100
 last_date <- Sys.Date()
 
