+++ NEWS.md
@@ -1,3 +1,8 @@
+## Version 1.0.1 (2022-08-15) -- bug fixex
+
+- fixed bug with FB ticker (change to TSLA)
+
+
 ## Version 1.0.0 (2022-06-22) -- CRAN SUBMISSION
 
 - many pkg changes after [ropensci approval](https://github.com/ropensci/software-review/issues/523)
