+++ DESCRIPTION
@@ -36,7 +36,7 @@ Imports:
     methods
 License: MIT + file LICENSE
 LazyData: true
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 Suggests: 
     knitr,
     rmarkdown,
