### 2022-08-25 3d479bb
fixed ticker FB to META
